refactor(desk98): author the Vue Vapor desktop in JSX instead of SFCs
Same framework, different authoring surface: the 18 .vue single-file
components collapse back into the original three-module shape — chrome.tsx
(T98, caption controls, taskbar, popups, Start menu, icons), programs.tsx
(program views + content-local hit helpers back in one file) and app.tsx
(the compositor, with the window shell + content dispatch) — compiled
through vue-jsx-vapor like apps/hero-vue-vapor and the app.vue-vapor.tsx
variants. Reactivity, state modules, the desk dialect, tests and the build
wiring are untouched; pocket.json only moves the entry to main.tsx.

JSX-specific rules honored: lists are plain .map() (the repo's vapor JSX
idiom), reactive branches live in JSX child positions (a bare ternary
returned from a component body evaluates once — MinesCell keeps its
hidden/revealed swap inside a stable 16×16 wrapper), and imports use the
clean @pocketjs/framework/* subpaths the framework resolves per-framework.
