+++ apps/desk98/main.tsx
@@ -1,0 +1,5 @@
+// apps/desk98/main.tsx — PocketJS 98 entry point (Vue Vapor, JSX).
+import { mount } from "@pocketjs/framework";
+import App from "./app.tsx";
+
+mount(() => <App />);
